Ignore messages not meant for the bot
The bot will now only respond to messages that are either posted in a
private channel to him or are prefixed with the bot command (.sh).
Messages in private channels must not prefixed with it (or it'll
complain that it doesn't know the command ".sh"). Might be a good idea
to make the bot command optional in private chat for usability.