Shut down via command line
The only way to gracefully shut down the bot is now by entering "s" or
"shutdown" on the command line. The "shutdown" command to kill it on
discord has been removed.

There is a separate thread listening to stdin, which sends a message to
the main thread once "shutdown" was entered. Unfortunately, the main
thread spends most of its time being blocked by
discord::Connection::recv_event(), so it can only react to that after it
receives and processes the event it's waiting for.