Add rewind to tokenizer
SplitWhitespaceWithRest can now rewind by one token, i.e. when calling
next(), rewind() and next() again, the two calls to next() will return
the same value as long as the first one didn't return None. If the first
one returned None (i.e. the entire string was consumed), the iterator is
rewound so that the last token of the string is returned again.