Change data storage
Data about users (which tier they want to play) is now stored as a set
of UserDatum objects, each of which contains one tier. If a user wants
to play any tier, instead of three true flags, there will be three
UserDatum objects with the different tiers.

This will allow to set preferences independent of tiers as well as
multiple preferences for the same tier. E.g., a user could say he's
always up for tier 10, but can also say he's up for other tiers
temporarily.