Implement serialization
Implements rustc_serialize's Encodable and Decodable for ShStatus and
everything it contains so that it can be persisted to disk.

Unfortunately, Timeframe has to be encoded as a string (instead of as a
struct). It has to be used as a key in a map and anything other than
primitives don't work with the json encoder that I'm testing the
implementation with.